Include encapsulation randomness in test vectors
It wasn't obvious what to use for EncapDerand of DHKEM.

RFC 9180 uses rejection sampling for DeriveKeyPair, but includes a
scalar in the test vectors (which is not really an EncapDerand input).

draft-irtf-cfrg-hybrid-kems defines the EncapsDerand of a nominal group
to be reduction based.

Since DHKEM already has a DeriveKeyPair, I went for that.

Fixes #29